# src/api/utils/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API ",
    "description": "Essa API tem como objetivo controlar a cadeira e registrar testes",
    "termsOfService": "http://google.com",
    "contact": {
      "email": "gregory.lira@gmail.com"
    }
  },
  "version": "1.0.0",
  "servers": [
    {
      "url": "https://uploadgallery.herokuapp.com"
    }
  ],

  "apis": ["../src/api/index.ts"],
  "paths": {
    "/api/posts": {
      "post": {
        "summary": "postagem de image",
        "description": "essa rota sera responsavel por postar as imagens",
        "tags": ["Image"],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Image"
              },
              "examples": {
                "tag": {
                  "value": {
                    "email": "ademir@ademir.com",
                    "password": "12345"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid tag post"
          },
          "201": {
            "description": "OK",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "get": {
        "summary": "Listagem de imagem",
        "description": "essa rota sera responsavel por listar todas as imagem",
        "tags": ["Image"],

        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid tag post"
          },
          "201": {
            "description": "OK",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "delete": {
        "summary": "Deletar imagem por id",
        "description": "essa rota sera responsavel por deletar imagens",
        "tags": ["Imagem"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id da imagem",
            "require": true
          }
        ],
        "responses": {
          "400": {
            "description": "Imagem não foi encontrada"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Image": {
        "type": "object",
        "properties": {
          "file": {
            "type": "file"
          },
          "tag": {
            "type": "string"
          }
        }
      }
    }
  }
}